+++ packages/npm/@amazeelabs/react-framework-bridge/src/types.ts
@@ -9,13 +9,17 @@ export type Html = React.VFC<{
   classNames?: {
     [key: string]: string | ClassFunction;
   };
-}>;
+}> & {
+  initialHtmlString: string;
+};
 
 export type HtmlBuilder = (input: string) => Html;
 
 export type Image = React.VFC<{
   className?: string;
-}>;
+}> & {
+  src: string | undefined;
+};
 
 export type ImageProps = Omit<
   React.DetailedHTMLProps<
@@ -35,6 +39,7 @@ export type Link<Query extends Parameters<typeof stringify>[0] = {}> =
     fragment?: string;
   }> & {
     navigate: (opts?: { query?: Query; fragment?: string }) => void;
+    href: string;
   };
 
 export type LinkProps<Query extends Parameters<typeof stringify>[0] = {}> =