+++ packages/npm/@silverback/ui/.storybook/preview.js
@@ -1,0 +1,16 @@
+import '../tailwind.css';
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+  options: {
+    storySort: {
+      order: [
+        'Atoms',
+        'Molecules',
+        'Organisms',
+        'Layouts',
+        'Pages'
+        ['Atoms', 'Molecules', 'Organisms', 'Pages', 'Layouts'],
+      ],
+    },
+  },
+};