+++ apps/silverback-gatsby/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.8](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gastby@1.2.7...@-amazeelabs/silverback-gastby@1.2.8) (2021-10-04)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gastby
+
+
+
+
+
+## [1.2.7](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gastby@1.2.6...@-amazeelabs/silverback-gastby@1.2.7) (2021-09-28)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gastby
+
+
+
+
+
+## [1.2.6](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gastby@1.2.5...@-amazeelabs/silverback-gastby@1.2.6) (2021-09-28)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gastby
+
+
+
+
+
+## [1.2.5](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gastby@1.2.4...@-amazeelabs/silverback-gastby@1.2.5) (2021-09-28)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gastby
+
+
+
+
+
 ## [1.2.4](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gastby@1.2.3...@-amazeelabs/silverback-gastby@1.2.4) (2021-09-20)
 
 **Note:** Version bump only for package @-amazeelabs/silverback-gastby