+++ packages/npm/@amazeelabs/gatsby-theme-core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.18](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-theme-core@0.6.17...@amazeelabs/gatsby-theme-core@0.6.18) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-theme-core
+
+
+
+
+
+## [0.6.17](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-theme-core@0.6.16...@amazeelabs/gatsby-theme-core@0.6.17) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-theme-core
+
+
+
+
+
+## [0.6.16](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-theme-core@0.6.15...@amazeelabs/gatsby-theme-core@0.6.16) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/gatsby-theme-core
+
+
+
+
+
 ## [0.6.15](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/gatsby-theme-core@0.6.14...@amazeelabs/gatsby-theme-core@0.6.15) (2021-09-15)
 
 **Note:** Version bump only for package @amazeelabs/gatsby-theme-core