+++ packages/npm/@amazeelabs/react-di/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.30](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.29...@amazeelabs/react-di@1.0.30) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
+## [1.0.29](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.28...@amazeelabs/react-di@1.0.29) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
+## [1.0.28](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.27...@amazeelabs/react-di@1.0.28) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
 ## [1.0.27](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.26...@amazeelabs/react-di@1.0.27) (2021-09-15)
 
 **Note:** Version bump only for package @amazeelabs/react-di