+++ packages/npm/@amazeelabs/silverback-playwright/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.2...@amazeelabs/silverback-playwright@1.3.3) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.3.2](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.1...@amazeelabs/silverback-playwright@1.3.2) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.3.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.0...@amazeelabs/silverback-playwright@1.3.1) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
 # [1.3.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.2.2...@amazeelabs/silverback-playwright@1.3.0) (2021-09-15)