+++ packages/tests/silverback-iframe/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-iframe-tests@1.0.11...@-amazeelabs/silverback-iframe-tests@1.0.12) (2021-10-04)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-iframe-tests
+
+
+
+
+
+## [1.0.11](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-iframe-tests@1.0.10...@-amazeelabs/silverback-iframe-tests@1.0.11) (2021-09-28)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-iframe-tests
+
+
+
+
+
+## [1.0.10](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-iframe-tests@1.0.9...@-amazeelabs/silverback-iframe-tests@1.0.10) (2021-09-28)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-iframe-tests
+
+
+
+
+
 ## [1.0.9](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-iframe-tests@1.0.8...@-amazeelabs/silverback-iframe-tests@1.0.9) (2021-09-15)
 
 **Note:** Version bump only for package @-amazeelabs/silverback-iframe-tests